GRHayL_IGH: Removed unneeded variables; simplified scheduling
BSSN variables aren't needed for the pure hydro code, and many of
the temp variables (reconstruction and characteristic speeds)
aren't needed because there's no second reconstruction due to A_i.
By combining those functions into a single loop, we remove the grid
functions and remove several function calls
