+++ implementations/GRHayL_IGH/src/enforce_detgtij_eq_1.c
@@ -1,15 +1,8 @@
-#include "cctk.h"
-#include "cctk_Parameters.h"
 #include "IGH.h"
 
-void GRHayL_IGH_convert_ADM_to_BSSN(const cGH *cctkGH,
-      double *gxx, double *gxy, double *gxz,
-      double *gyy, double *gyz, double *gzz,
-      double *phi, double *psi,
-      double *gtxx, double *gtxy, double *gtxz,
-      double *gtyy, double *gtyz, double *gtzz,
-      double *gtupxx, double *gtupxy, double *gtupxz,
-      double *gtupyy, double *gtupyz, double *gtupzz) {
+void GRHayL_IGH_enforce_detgtij_eq_1(CCTK_ARGUMENTS) {
+  DECLARE_CCTK_ARGUMENTS_GRHayL_IGH_enforce_detgtij_eq_1;
+  DECLARE_CCTK_PARAMETERS;
 
 #pragma omp parallel for
   for(int k=0; k<cctkGH->cctk_lsh[2]; k++)
@@ -31,8 +24,7 @@ void GRHayL_IGH_convert_ADM_to_BSSN(const cGH *cctkGH,
 
         gijdet = fabs(gijdet);
 
-        const double phiL = (1.0/12.0) * log(gijdet);
-        const double psiL = exp(phiL);
+        const double psiL = exp( (1.0/12.0) * log(gijdet) );
 
         const double Psim4 = 1.0/(psiL*psiL*psiL*psiL);
         double gtxxL = gxx_physL*Psim4;
@@ -61,32 +53,15 @@ void GRHayL_IGH_convert_ADM_to_BSSN(const cGH *cctkGH,
                                      "WARNING: det[3-metric]<0.0 at point  %d %d %d | cctk_lsh: %d %d %d. Hopefully this is occurring in gz's! gtij_phys = %.2e %.2e %.2e %.2e %.2e %.2e gtij_new = %.2e %.2e %.2e %.2e %.2e %.2e | gijdet = %.2e | gtijdet = %.2e",
 				     i,j,k,cctkGH->cctk_lsh[0],cctkGH->cctk_lsh[1],cctkGH->cctk_lsh[2],gxx_physL,gxy_physL,gxz_physL,gyy_physL,gyz_physL,gzz_physL,gtxxL,gtxyL,gtxzL,gtyyL,gtyzL,gtzzL,-gijdet,gtijdet); }
 
+        /*******************************************
+         * Set the ADM gridfunctions to new values *
+         *******************************************/
         const double Psi4 = psiL*psiL*psiL*psiL;
-        /*****************************************
-         * Set all the needed BSSN gridfunctions *
-         *****************************************/
-        phi[index] = phiL;
-        psi[index] = psiL;
-
-        gtxx[index] = gtxxL;
-        gtxy[index] = gtxyL;
-        gtxz[index] = gtxzL;
-        gtyy[index] = gtyyL;
-        gtyz[index] = gtyzL;
-        gtzz[index] = gtzzL;
-
         gxx[index] = gtxxL*Psi4;
         gxy[index] = gtxyL*Psi4;
         gxz[index] = gtxzL*Psi4;
         gyy[index] = gtyyL*Psi4;
         gyz[index] = gtyzL*Psi4;
         gzz[index] = gtzzL*Psi4;
-
-        gtupxx[index] =   ( gtyyL * gtzzL - gtyzL * gtyzL );
-        gtupxy[index] = - ( gtxyL * gtzzL - gtyzL * gtxzL );
-        gtupxz[index] =   ( gtxyL * gtyzL - gtyyL * gtxzL );
-        gtupyy[index] =   ( gtxxL * gtzzL - gtxzL * gtxzL );
-        gtupyz[index] = - ( gtxxL * gtyzL - gtxyL * gtxzL );
-        gtupzz[index] =   ( gtxxL * gtyyL - gtxyL * gtxyL );
   }
 }
