+++ implementations/GRHayL_IGH/src/make.code.defn
@@ -1,16 +1,14 @@
 # Main make.code.defn file for thorn GRHayL
 
 # Source files in this directory
-SRCS = calculate_HD_rhs.c                             \
-       compute_characteristic_speeds.c                \
-       conservs_to_prims.c                            \
-       convert_ADM_to_BSSN.c                          \
-       convert_from_ADMBase_HydroBase_to_GRHayL_IGH.c \
-       convert_from_GRHayL_IGH_to_HydroBase.c         \
-       evaluate_HD_rhs.c                              \
-       InitSymBound.c                                 \
-       interpolate_to_face_and_initialize_metric.c    \
-       MoL_registration.c                             \
-       outer_boundaries.c                             \
-       postpostinitial__set_symmetries__copy_timelevels.c \
+SRCS = conservs_to_prims.c                         \
+       enforce_detgtij_eq_1.c                      \
+       convert_HydroBase_to_GRHayL_IGH.c           \
+       convert_from_GRHayL_IGH_to_HydroBase.c      \
+       evaluate_flux_source_rhs.c                  \
+       evaluate_tau_curvature_rhs.c                \
+       InitSymBound.c                              \
+       interpolate_to_face_and_initialize_metric.c \
+       MoL_registration.c                          \
+       outer_boundaries.c                          \
        reconstruction_loop.c
